# Test custom PCG performance