--- a/benchmark/benchmark_hydrostatic_model.jl
+++ b/benchmark/benchmark_hydrostatic_model.jl
@@ -3,16 +3,14 @@ push!(LOAD_PATH, joinpath(@__DIR__, ".."))
 using Benchmarks
 
 using Oceananigans.TimeSteppers: update_state!
-using Oceananigans.Diagnostics: accurate_cell_advection_timescale
-
 using BenchmarkTools
 using CUDA
 using Oceananigans
 using Statistics
 
 # Problem size
-Nx = 256
-Ny = 128
+Nx = 512
+Ny = 256
 
 function set_divergent_velocity!(model)
     # Create a divergent velocity
@@ -34,39 +32,28 @@ function set_divergent_velocity!(model)
     return nothing
 end
 
+λ¹ₚ = 70
+λ²ₚ = 70 + 180
+φₚ  = 55
+
+# We need a bottom height field that ``masks'' the singularities
+bottom_height(λ, φ) = ((abs(λ - λ¹ₚ) < 5) & (abs(φₚ - φ) < 5)) |
+                      ((abs(λ - λ²ₚ) < 5) & (abs(φₚ - φ) < 5)) | (φ < -78) ? 1 : 0
+
 grids = Dict(
     (CPU, :RectilinearGrid)       => RectilinearGrid(CPU(), size=(Nx, Ny, 1), extent=(1, 1, 1)),
     (CPU, :LatitudeLongitudeGrid) => LatitudeLongitudeGrid(CPU(), size=(Nx, Ny, 1), longitude=(-180, 180), latitude=(-80, 80), z=(-1, 0), precompute_metrics=true),
+    (CPU, :TripolarGrid)          => ImmersedBoundaryGrid(TripolarGrid(CPU(), size=(Nx, Ny, 1)), GridFittedBottom(bottom_height)), 
     (GPU, :RectilinearGrid)       => RectilinearGrid(GPU(), size=(Nx, Ny, 1), extent=(1, 1, 1)),
     (GPU, :LatitudeLongitudeGrid) => LatitudeLongitudeGrid(GPU(), size=(Nx, Ny, 1), longitude=(-160, 160), latitude=(-80, 80), z=(-1, 0), precompute_metrics=true),
+    (GPU, :TripolarGrid)          => ImmersedBoundaryGrid(TripolarGrid(GPU(), size=(Nx, Ny, 1)), GridFittedBottom(bottom_height)) 
 )
-
-# Cubed sphere cases maybe worth considering eventually
-#=
-using DataDeps
-
-ENV["DATADEPS_ALWAYS_ACCEPT"] = "true"
-
-dd = DataDep("cubed_sphere_510_grid",
-    "Conformal cubed sphere grid with 510×510 grid points on each face",
-    "https://engaging-web.mit.edu/~alir/cubed_sphere_grids/cs510/cubed_sphere_510_grid.jld2"
-)
-
-DataDeps.register(dd)
-
-# (CPU, :ConformalCubedSphereGrid)     => ConformalCubedSphereGrid(datadep"cubed_sphere_510_grid/cubed_sphere_510_grid.jld2", Nz=1, z=(-1, 0)),
-# (GPU, :ConformalCubedSphereGrid)     => ConformalCubedSphereGrid(datadep"cubed_sphere_510_grid/cubed_sphere_510_grid.jld2", Nz=1, z=(-1, 0), architecture=GPU()),
-=#
 
 free_surfaces = Dict(
     :ExplicitFreeSurface => ExplicitFreeSurface(),
+    :SplitExplicitFreeSurface => SplitExplicitFreeSurface(; substeps=50),
     :PCGImplicitFreeSurface => ImplicitFreeSurface(solver_method = :PreconditionedConjugateGradient), 
-    #:PCGImplicitFreeSurfaceNoPreconditioner => ImplicitFreeSurface(solver_method = :PreconditionedConjugateGradient, preconditioner_method = nothing), 
-    :MatrixImplicitFreeSurfaceOrd2 => ImplicitFreeSurface(solver_method = :HeptadiagonalIterativeSolver), 
-    #:MatrixImplicitFreeSurfaceOrd1 => ImplicitFreeSurface(solver_method = :HeptadiagonalIterativeSolver, preconditioner_settings= (order = 1,) ), 
-    #:MatrixImplicitFreeSurfaceOrd0 => ImplicitFreeSurface(solver_method = :HeptadiagonalIterativeSolver, preconditioner_settings= (order = 0,) ), 
-    #:MatrixImplicitFreeSurfaceNoPreconditioner => ImplicitFreeSurface(solver_method = :HeptadiagonalIterativeSolver, preconditioner_method = nothing),
-    :MatrixImplicitFreeSurfaceSparsePreconditioner => ImplicitFreeSurface(solver_method = :HeptadiagonalIterativeSolver, preconditioner_method = :SparseInverse)
+    :MatrixImplicitFreeSurface => ImplicitFreeSurface(solver_method = :HeptadiagonalIterativeSolver), 
 )
 
 function benchmark_hydrostatic_model(Arch, grid_type, free_surface_type)
@@ -74,11 +61,11 @@ function benchmark_hydrostatic_model(Arch, grid_type, free_surface_type)
     grid = grids[(Arch, grid_type)]
 
     model = HydrostaticFreeSurfaceModel(; grid,
-                                        momentum_advection = VectorInvariant(),
-                                        free_surface = free_surfaces[free_surface_type])
+                                          momentum_advection = VectorInvariant(),
+                                          free_surface = free_surfaces[free_surface_type])
 
     set_divergent_velocity!(model)
-    Δt = accurate_cell_advection_timescale(grid, model.velocities) / 2
+    Δt = Oceananigans.Advection.cell_advection_timescale(grid, model.velocities) / 2
     time_step!(model, Δt) # warmup
     
     trial = @benchmark begin
@@ -90,15 +77,12 @@ end
 
 # Benchmark parameters
 
-#architectures = has_cuda() ? [GPU] : [CPU]
-architectures = [CPU]
+architectures = has_cuda() ? [GPU, CPU] : [CPU]
 
 grid_types = [
     :RectilinearGrid,
     :LatitudeLongitudeGrid,
-    # Uncomment when OrthogonalSphericalShellGrids of any size can be built natively without loading from file:
-    # :OrthogonalSphericalShellGrid,
-    # :ConformalCubedSphereGrid
+    :TripolarGrid
 ]
 
 free_surface_types = collect(keys(free_surfaces))