--- a/benchmark/benchmark_hydrostatic_model.jl
+++ b/benchmark/benchmark_hydrostatic_model.jl
@@ -51,8 +51,8 @@ grids = Dict(
 )
 
 free_surfaces = Dict(
-   :ExplicitFreeSurface => ExplicitFreeSurface(),
-   :SplitExplicitFreeSurface => SplitExplicitFreeSurface(; substeps=50),
+#    :ExplicitFreeSurface => ExplicitFreeSurface(),
+#    :SplitExplicitFreeSurface => SplitExplicitFreeSurface(; substeps=50),
    :KrylovImplicitFreeSurface => ImplicitFreeSurface(solver_method=:PreconditionedConjugateGradient, preconditioner=nothing, Solver=KrylovSolver), 
    :PCGImplicitFreeSurface    => ImplicitFreeSurface(solver_method=:PreconditionedConjugateGradient), 
    :MatrixImplicitFreeSurface => ImplicitFreeSurface(solver_method=:HeptadiagonalIterativeSolver),