Oceananigans' custom PCG has a couple of performance problems. If the objective is eventually to remove the `Matrix` solver we need to make sure that the PCG is up to speed. This PR is propedeutic to #4386 and refurbishes some benchmark tests that we can use to improve the PCG performance. The result at the moment is that the PCG works well only on rectilinear non-immersed grids.

Another PR will come before #4386 to bring the PCG up to speed

The output of the benchmark is here
```julia
32×10 DataFrame
 Row │ architectures  grid_types             free_surface_types         min         median      mean        max         memory      allocs  samples
     │ Any            Any                    Any                        Any         Any         Any         Any         Any         Any     Any
─────┼──────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────
   1 │ CPU            ImmersedLatGrid        ExplicitFreeSurface        50.997 ms   51.075 ms   51.320 ms   52.579 ms   1.11 MiB    1070    10
   2 │ CPU            ImmersedLatGrid        MatrixImplicitFreeSurface  132.211 ms  142.570 ms  143.763 ms  162.111 ms  1.99 MiB    3065    10
   3 │ CPU            ImmersedLatGrid        PCGImplicitFreeSurface     373.622 ms  391.803 ms  398.345 ms  438.238 ms  136.70 MiB  35528   10
   4 │ CPU            ImmersedLatGrid        SplitExplicitFreeSurface   179.234 ms  235.579 ms  238.157 ms  309.451 ms  5.22 MiB    4323    10
   5 │ CPU            ImmersedRecGrid        ExplicitFreeSurface        41.571 ms   42.001 ms   42.036 ms   42.755 ms   746.29 KiB  1027    10
   6 │ CPU            ImmersedRecGrid        MatrixImplicitFreeSurface  142.778 ms  144.371 ms  144.757 ms  148.521 ms  1.31 MiB    3017    10
   7 │ CPU            ImmersedRecGrid        PCGImplicitFreeSurface     187.274 ms  190.594 ms  194.484 ms  215.125 ms  61.00 MiB   17563   10
   8 │ CPU            ImmersedRecGrid        SplitExplicitFreeSurface   293.361 ms  354.231 ms  352.831 ms  406.777 ms  3.78 MiB    4349    10
   9 │ CPU            LatitudeLongitudeGrid  ExplicitFreeSurface        27.922 ms   28.256 ms   28.422 ms   29.354 ms   562.52 KiB  1006    10
  10 │ CPU            LatitudeLongitudeGrid  MatrixImplicitFreeSurface  93.440 ms   94.291 ms   96.596 ms   113.497 ms  1.00 MiB    2937    10
  11 │ CPU            LatitudeLongitudeGrid  PCGImplicitFreeSurface     259.576 ms  277.398 ms  280.766 ms  301.446 ms  130.83 MiB  35400   10
  12 │ CPU            LatitudeLongitudeGrid  SplitExplicitFreeSurface   114.787 ms  119.239 ms  119.896 ms  126.970 ms  2.61 MiB    4170    10
  13 │ CPU            RectilinearGrid        ExplicitFreeSurface        24.716 ms   25.500 ms   25.504 ms   26.809 ms   374.48 KiB  963     10
  14 │ CPU            RectilinearGrid        MatrixImplicitFreeSurface  73.609 ms   80.298 ms   80.621 ms   89.221 ms   696.18 KiB  2885    10
  15 │ CPU            RectilinearGrid        PCGImplicitFreeSurface     40.315 ms   41.191 ms   41.080 ms   41.950 ms   2.85 MiB    3610    10
  16 │ CPU            RectilinearGrid        SplitExplicitFreeSurface   206.162 ms  228.030 ms  225.412 ms  240.865 ms  1.93 MiB    4196    10
  17 │ GPU            ImmersedLatGrid        ExplicitFreeSurface        50.364 ms   51.378 ms   51.420 ms   52.472 ms   1.11 MiB    1070    10
  18 │ GPU            ImmersedLatGrid        MatrixImplicitFreeSurface  115.014 ms  116.209 ms  118.472 ms  124.939 ms  1.99 MiB    3065    10
  19 │ GPU            ImmersedLatGrid        PCGImplicitFreeSurface     372.160 ms  392.106 ms  391.975 ms  410.496 ms  136.70 MiB  35530   10
  20 │ GPU            ImmersedLatGrid        SplitExplicitFreeSurface   199.555 ms  281.901 ms  270.142 ms  341.260 ms  5.22 MiB    4323    10
  21 │ GPU            ImmersedRecGrid        ExplicitFreeSurface        918.828 μs  960.679 μs  1.022 ms    1.530 ms    905.59 KiB  2112    10
  22 │ GPU            ImmersedRecGrid        MatrixImplicitFreeSurface  12.784 ms   12.935 ms   12.970 ms   13.494 ms   1.88 MiB    20020   10
  23 │ GPU            ImmersedRecGrid        PCGImplicitFreeSurface     24.841 ms   25.105 ms   25.156 ms   26.013 ms   6.65 MiB    47592   10
  24 │ GPU            ImmersedRecGrid        SplitExplicitFreeSurface   3.221 ms    3.473 ms    3.500 ms    3.958 ms    2.43 MiB    9096    10
  25 │ GPU            LatitudeLongitudeGrid  ExplicitFreeSurface        1.016 ms    1.058 ms    1.130 ms    1.701 ms    701.93 KiB  2861    10
  26 │ GPU            LatitudeLongitudeGrid  MatrixImplicitFreeSurface  19.470 ms   22.915 ms   22.875 ms   27.270 ms   1.79 MiB    31199   10
  27 │ GPU            LatitudeLongitudeGrid  PCGImplicitFreeSurface     100.270 ms  100.795 ms  105.360 ms  141.808 ms  16.34 MiB   183968  10
  28 │ GPU            LatitudeLongitudeGrid  SplitExplicitFreeSurface   3.708 ms    4.011 ms    4.010 ms    4.719 ms    2.17 MiB    10043   10
  29 │ GPU            RectilinearGrid        ExplicitFreeSurface        717.393 μs  732.261 μs  797.732 μs  1.331 ms    458.16 KiB  1849    10
  30 │ GPU            RectilinearGrid        MatrixImplicitFreeSurface  11.960 ms   12.040 ms   12.073 ms   12.491 ms   1.15 MiB    19490   10
  31 │ GPU            RectilinearGrid        PCGImplicitFreeSurface     3.365 ms    3.414 ms    3.574 ms    4.844 ms    1.04 MiB    7038    10
  32 │ GPU            RectilinearGrid        SplitExplicitFreeSurface   2.275 ms    2.302 ms    2.390 ms    3.014 ms    1.50 MiB    8542    10
```