@glwagner 2025-04-11
@amontoison @simone-silvestri the objective is to use Krylove solvers, not Oceananigans PCG right?