@glwagner 2025-04-11
Also, is this work worthwhile if we don't want to use the matrix solver? The maintenance and testing has a cost, I am wondering whether it is worth it.