@amontoison 2025-04-11
Yes, the goal is to use the solvers in Krylov.jl, which should not be less efficient than your PCG here or the one in IterativeSolvers.jl.