@amontoison 2025-04-12
@simone-silvestri The `KrylovSolver` is now available in `main`, is it easy to try it and see the difference with the current version?