@simone-silvestri 2025-04-15
@amontoison I think I need your help here. With the Krylov solver, the benchmarks error with 
```julia
[2025/04/15 05:05:38.140] INFO  Benchmarking 1/8: (GPU, :RectilinearGrid, :KrylovImplicitFreeSurface)...
ERROR: LoadError: DomainError with -0.0506264973347019:
sqrt was called with a negative real argument but will only return a complex result if called with a complex argument. Try sqrt(Complex(x)).
Stacktrace:
  [1] throw_complex_domainerror(f::Symbol, x::Float64)
    @ Base.Math ./math.jl:33
  [2] sqrt(x::Float64)
    @ Base.Math ./math.jl:686
  [3] cg!(solver::Krylov.CgSolver{…}, A::Oceananigans.Solvers.KrylovOperator{…}, b::Oceananigans.Solvers.KrylovField{…}; M::Oceananigans.Solvers.KrylovPreconditioner{…}, ldiv::Bool, radius::Float64, linesearch::Bool, atol::Float64, rtol::Float64, itmax::Int64, timemax::Float64, verbose::Int64, history::Bool, callback::Krylov.var"#837#845", iostream::Core.CoreSTDOUT)
    @ Krylov ~/.julia/packages/Krylov/wgV9p/src/cg.jl:146
  [4] cg!
    @ ~/.julia/packages/Krylov/wgV9p/src/cg.jl:108 [inlined]
  [5] solve!
    @ ~/.julia/packages/Krylov/wgV9p/src/krylov_solve.jl:159 [inlined]
  [6] solve!(::Field{…}, ::KrylovSolver{…}, ::Field{…}, ::Field{…}, ::Vararg{…}; kwargs::@Kwargs{})
    @ Oceananigans.Solvers ~/development/Oceananigans.jl/src/Solvers/krylov_solver.jl:150
  [7] solve!
    @ ~/development/Oceananigans.jl/src/Solvers/krylov_solver.jl:147 [inlined]
  [8] solve!
    @ ~/development/Oceananigans.jl/src/Models/HydrostaticFreeSurfaceModels/pcg_implicit_free_surface_solver.jl:95 [inlined]
  [9] step_free_surface!(free_surface::ImplicitFreeSurface{…}, model::HydrostaticFreeSurfaceModel{…}, timestepper::Oceananigans.TimeSteppers.QuasiAdamsBashforth2TimeStepper{…}, Δt::Float64)
    @ Oceananigans.Models.HydrostaticFreeSurfaceModels ~/development/Oceananigans.jl/src/Models/HydrostaticFreeSurfaceModels/implicit_free_surface.jl:145
 [10] ab2_step!
    @ ~/development/Oceananigans.jl/src/Models/HydrostaticFreeSurfaceModels/hydrostatic_free_surface_ab2_step.jl:24 [inlined]
 [11] time_step!(model::HydrostaticFreeSurfaceModel{…}, Δt::Float64; callbacks::Vector{…}, euler::Bool)
    @ Oceananigans.TimeSteppers ~/development/Oceananigans.jl/src/TimeSteppers/quasi_adams_bashforth_2.jl:99
 [12] time_step!(model::HydrostaticFreeSurfaceModel{…}, Δt::Float64)
    @ Oceananigans.TimeSteppers ~/development/Oceananigans.jl/src/TimeSteppers/quasi_adams_bashforth_2.jl:74
 [13] benchmark_hydrostatic_model(Arch::Type, grid_type::Symbol, free_surface_type::Symbol)
    @ Main ~/development/Oceananigans.jl/benchmark/benchmark_hydrostatic_model.jl:70
 [14] run_benchmarks(benchmark_fun::Function; kwargs::@Kwargs{…})
    @ Benchmarks ~/development/Oceananigans.jl/benchmark/src/Benchmarks.jl:57
 [15] top-level scope
    @ ~/development/Oceananigans.jl/benchmark/benchmark_hydrostatic_model.jl:94
 [16] include(fname::String)
    @ Base.MainInclude ./client.jl:494
 [17] top-level scope
    @ REPL[1]:1
in expression starting at /home/ssilvest/development/Oceananigans.jl/benchmark/benchmark_hydrostatic_model.jl:94
Some type information was truncated. Use `show(err)` to see complete types.
```
Do I need to change something with respect to the custom PCG to use the Krylov solver?