@simone-silvestri 2025-04-15
I think the preconditioner acts differently with a `KrylovSolver` with respect to our custom `ConjugateGradientSolver`. The negative sqrt error appears only using a preconditioner.